requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.25.0
asyncpg==0.29.0 